Moved appropriate fields, getters and setters, etc to AbstractEntity, made Item an extension of AE.
