+++ back-end/src/main/java/org/launchcode/backend/models/AbstractEntity.java
@@ -1,7 +1,51 @@
 package org.launchcode.backend.models;
 
-//@MappedSuperclass
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
+import java.util.Objects;
+
+@MappedSuperclass
 public abstract class AbstractEntity {
 
+    @Id
+    @GeneratedValue
+    public int id; //changed fields temporarily to public until we resolve what to do between FE/BE
 
+    @NotBlank(message="Title is required.")
+    @Size(min=3, max=100, message="Name must be between 3 and 100 characters.")
+    public String title;
+
+    public AbstractEntity(String title) {
+        this.title = title;
+    }
+
+    public AbstractEntity() {}
+
+    public int getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        AbstractEntity that = (AbstractEntity) o;
+        return id == that.id;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
+    }
 }
